Rename project to `django-hcaptcha-field`
This avoids a clash with the base project on GitHub and PyPi. While at
it, the README file has been converted to Markdown for easier editing.
