+++ MANIFEST.in
@@ -1,4 +1,4 @@
 include LICENSE
 include README.rst
-recursive-include hcaptcha/templates *
+recursive-include hcaptcha_field/templates *
 recursive-include docs *
