+++ hcaptcha_field/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class hCaptchaFieldConfig(AppConfig):
+    name = 'hcaptcha_field'
+    verbose_name = 'hCaptcha Field'
