+++ setup.cfg
@@ -1,11 +1,11 @@
 [metadata]
-name = django-hCaptcha
+name = django-hCaptcha-field
 version = 0.1.0
-description = Django hCaptcha provides a simple way to protect your django forms using hCaptcha
+description = Django hCaptcha Field provides a simple way to protect your Django forms using hCaptcha.
 long_description = file: README.rst
-url = https://github.com/AndrejZbin
-author = Andrej Zbín
-author_email = zbin.andrej@gmail.com
+url = https://github.com/tiesjan/django-hcaptcha-field
+author = Ties Jan Hefting
+author_email = hello@tiesjan.com
 license = BSD-3-Clause
 classifiers =
     Environment :: Web Environment
